Re-organize recordings apis
